Moved TWGameManager::processMovement to its own class.
